Extract DocMethods::EntityParser
