+++ lib/grape-swagger.rb
@@ -2,6 +2,7 @@ require 'grape'
 require 'grape-swagger/version'
 require 'grape-swagger/errors'
 require 'grape-swagger/doc_methods'
+require 'grape-swagger/doc_methods/entity_parser'
 require 'grape-swagger/markdown'
 require 'grape-swagger/markdown/kramdown_adapter'
 require 'grape-swagger/markdown/redcarpet_adapter'
